Subscribe customer using data from edd_complete_purchase
I think it makes the most sense to consolidate all `edd_mailchimp` data
in one action hook rather than attempting to process subscription data
several different times in a manner that may not be idempotent.

This allows us to fetch all download data, choke it down to the basics of what we need
to make successful requests, and process all in one hook.